+++ Controller/PageEditController.php
@@ -17,15 +17,14 @@ use Symfony\Component\Filesystem\Filesystem;
 use Symfony\Component\Finder\Finder;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\HttpFoundation\Session\Session;
 use Symfony\Component\HttpKernel\Exception\HttpException;
-use Cocur\Slugify\Slugify;
 use c975L\PageEditBundle\Entity\PageEdit;
 use c975L\PageEditBundle\Form\PageEditType;
+use c975L\PageEditBundle\Service\PageEditService;
 
 class PageEditController extends Controller
 {
-    //DASHBOARD
+//DASHBOARD
     /**
      * @Route("/pages/dashboard",
      *      name="pageedit_dashboard")
@@ -39,7 +38,8 @@ class PageEditController extends Controller
         //Returns the dashboard content
         if ($user !== null && $this->get('security.authorization_checker')->isGranted($this->getParameter('c975_l_page_edit.roleNeeded'))) {
             //Creates structure in case it not exists
-            $this->createFolders();
+            $pageEditService = $this->get(PageEditService::class);
+            $pageEditService->createFolders();
 
             //Gets the Finder
             $finder = new Finder();
@@ -64,7 +64,7 @@ class PageEditController extends Controller
                 $slug = str_replace('.html.twig', '', $file->getRelativePathname());
                 preg_match('/pageedit_title=\"(.*)\"/', $file->getContents(), $matches);
                 if (!empty($matches)) {
-                    $title = $matches[1];
+                    $title = str_replace('\"', '"', $matches[1]);
                 } else {
                     //Title is using Twig code to translate it
                     preg_match('/pageedit_title=(.*)\%\}/', $file->getContents(), $matches);
@@ -188,10 +188,11 @@ class PageEditController extends Controller
 
             if ($form->isSubmitted() && $form->isValid()) {
                 //Gets slug
-                $slug = $this->slugify($form->getData()->getSlug());
+                $pageEditService = $this->get(PageEditService::class);
+                $slug = $pageEditService->slugify($form->getData()->getSlug());
 
                 //Writes file
-                $this->writeFile($slug, null, $form->getData(), $user->getId());
+                $pageEditService->writeFile($slug, null, $form->getData(), $user->getId());
 
                 //Redirects to the page
                 return $this->redirectToRoute('pageedit_display', array(
@@ -290,16 +291,17 @@ class PageEditController extends Controller
 
             if ($form->isSubmitted() && $form->isValid()) {
                 //Gets slug
-                $slug = $this->slugify($form->getData()->getSlug());
+                $pageEditService = $this->get(PageEditService::class);
+                $slug = $pageEditService->slugify($form->getData()->getSlug());
 
                 //Archives and redirects the file if title (then slug) has changed
                 if ($slug != $page) {
-                    $this->archiveFile($page, $user->getId());
-                    $this->redirectFile($page, $slug);
+                    $pageEditService->archiveFile($page, $user->getId());
+                    $pageEditService->redirectFile($page, $slug);
                 }
 
                 //Writes file
-                $this->writeFile($slug, $originalContent, $form->getData(), $user->getId());
+                $pageEditService->writeFile($slug, $originalContent, $form->getData(), $user->getId());
 
                 //Redirects to the page
                 return $this->redirectToRoute('pageedit_display', array(
@@ -384,7 +386,8 @@ class PageEditController extends Controller
 
             if ($form->isSubmitted() && $form->isValid()) {
                 //Deletes file
-                $this->deleteFile($page, false);
+                $pageEditService = $this->get(PageEditService::class);
+                $pageEditService->deleteFile($page, false);
 
                 //Redirects to the page which will be HTTP 410
                 return $this->redirectToRoute('pageedit_dashboard');
@@ -423,7 +426,8 @@ class PageEditController extends Controller
         //Returns the list content
         if ($user !== null && $this->get('security.authorization_checker')->isGranted($this->getParameter('c975_l_page_edit.roleNeeded'))) {
             //Creates structure in case it not exists
-            $this->createFolders();
+            $pageEditService = $this->get(PageEditService::class);
+            $pageEditService->createFolders();
 
             //Gets the Finder
             $finder = new Finder();
@@ -486,7 +490,8 @@ class PageEditController extends Controller
     public function uploadAction(Request $request, $page)
     {
         //Creates structure in case it not exists
-        $this->createFolders();
+        $pageEditService = $this->get(PageEditService::class);
+        $pageEditService->createFolders();
 
         //Defines path
         $folderPath = $this->getParameter('kernel.root_dir') . '/../web/images/' . $this->getParameter('c975_l_page_edit.folderPages');
@@ -522,7 +527,8 @@ class PageEditController extends Controller
      */
     public function slugAction($text)
     {
-        return $this->json(array('a' => $this->slugify($text)));
+        $pageEditService = $this->get(PageEditService::class);
+        return $this->json(array('a' => $pageEditService->slugify($text)));
     }
 
 //HELP
@@ -550,176 +556,5 @@ class PageEditController extends Controller
 
         //Access is denied
         throw $this->createAccessDeniedException();
-    }
-
-
-//FUNCTIONS
-    //Creates the folders needed by the Bundle
-    public function createFolders()
-    {
-        //Gets the FileSystem
-        $fs = new Filesystem();
-
-        //Defines paths
-        $folderPath = $this->getParameter('kernel.root_dir') . '/Resources/views/' . $this->getParameter('c975_l_page_edit.folderPages');
-        $archivedFolder = $folderPath . '/archived';
-        $deletedFolder = $folderPath . '/deleted';
-        $protectedFolderPath = $folderPath . '/protected';
-        $redirectedFolder = $folderPath . '/redirected';
-        $imageFolderPath = $this->getParameter('kernel.root_dir') . '/../web/images/' . $this->getParameter('c975_l_page_edit.folderPages');
-
-        //Creates folders
-        $fs->mkdir($archivedFolder, 0770);
-        $fs->mkdir($deletedFolder, 0770);
-        $fs->mkdir($protectedFolderPath, 0770);
-        $fs->mkdir($redirectedFolder, 0770);
-        $fs->mkdir($imageFolderPath, 0770);
-    }
-
-    //Gets the start and end of the skeleton
-    public function getSkeleton()
-    {
-        $parser = $this->get('templating.name_parser');
-        $locator = $this->get('templating.locator');
-
-        $skeleton = file_get_contents($locator->locate($parser->parse('c975LPageEditBundle::skeleton.html.twig')));
-
-        $startBlock = '{% block pageEdit %}';
-        $endBlock = '{% endblock %}';
-
-        $entryPoint = strpos($skeleton, $startBlock) + strlen($startBlock);
-        $exitPoint = strpos($skeleton, $endBlock, $entryPoint);
-
-        return array(
-            'startSkeleton' => trim(substr($skeleton, 0, $entryPoint)),
-            'endSkeleton' => trim(substr($skeleton, $exitPoint))
-        );
-    }
-
-
-    //Slugify function - https://github.com/cocur/slugify
-    public function slugify($text)
-    {
-        $slugify = new Slugify();
-        return $slugify->slugify($text);
-    }
-
-
-    //Archives file
-    public function archiveFile($page, $userId)
-    {
-        //Creates structure in case it not exists
-        $this->createFolders();
-
-        //Gets the FileSystem
-        $fs = new Filesystem();
-
-        //Defines paths
-        $folderPath = $this->getParameter('kernel.root_dir') . '/Resources/views/' . $this->getParameter('c975_l_page_edit.folderPages');
-        $archivedFolder = $folderPath . '/archived';
-        $filePath = $folderPath . '/' . $page . '.html.twig';
-
-        //Archives file
-        if ($fs->exists($filePath)) {
-            $fs->rename($filePath, $archivedFolder . '/' . $page . '-' . date('Ymd-His-') . $userId . '.html.twig');
-        }
-    }
-
-
-    //Creates the redirection file
-    public function redirectFile($page, $slug)
-    {
-        //Creates structure in case it not exists
-        $this->createFolders();
-
-        //Gets the FileSystem
-        $fs = new Filesystem();
-
-        //Defines path
-        $folderPath = $this->getParameter('kernel.root_dir') . '/Resources/views/' . $this->getParameter('c975_l_page_edit.folderPages');
-        $redirectedFolder = $folderPath . '/redirected';
-
-        //Sets the redirection
-        $redirectedFilePath = $redirectedFolder . '/' . $page . '.html.twig';
-        $fs->dumpFile($redirectedFilePath, $slug);
-    }
-
-
-    //Moves to deleted/redirected folder the requested file
-    public function deleteFile($page, $redirect, $slug = null)
-    {
-        //Creates structure in case it not exists
-        $this->createFolders();
-
-        //Gets the FileSystem
-        $fs = new Filesystem();
-
-        //Defines path
-        $folderPath = $this->getParameter('kernel.root_dir') . '/Resources/views/' . $this->getParameter('c975_l_page_edit.folderPages');
-        $filePath = $folderPath . '/' . $page . '.html.twig';
-        $deletedFolder = $folderPath . '/deleted';
-        $redirectedFolder = $folderPath . '/redirected';
-
-        //Sets the redirection
-        if ($redirect === true) {
-            $redirectedFilePath = $redirectedFolder . '/' . $page . '.html.twig';
-            $fs->dumpFile($redirectedFilePath, $slug);
-        }
-
-        //Deletes file
-        if ($fs->exists($filePath)) {
-            $fs->rename($filePath, $deletedFolder . '/' . $page . '.html.twig');
-        }
-    }
-
-
-    //Archives old file and writes new one
-    public function writeFile($page, $originalContent, $formData, $userId)
-    {
-        //Creates structure in case it not exists
-        $this->createFolders();
-
-        //Gets the FileSystem
-        $fs = new Filesystem();
-
-        //Defines path
-        $folderPath = $this->getParameter('kernel.root_dir') . '/Resources/views/' . $this->getParameter('c975_l_page_edit.folderPages');
-        $filePath = $folderPath . '/' . $page . '.html.twig';
-
-        //Gets the skeleton
-        extract($this->getSkeleton());
-
-        //Gets title
-        $title = $formData->getTitle();
-
-        //Title is using Twig code to translate it
-        if (strpos($title, '{{') === 0) {
-            $title = trim(str_replace(array('{{', '}}'), '', $title));
-        }
-        //Title is text
-        else {
-            $title = '"' . $title . '"';
-        }
-
-        //Updates metadata
-        $startSkeleton = preg_replace('/pageedit_title=\"(.*)\"/', 'pageedit_title=' . $title, $startSkeleton);
-        $startSkeleton = preg_replace('/pageedit_changeFrequency=\"(.*)\"/', 'pageedit_changeFrequency="' . $formData->getChangeFrequency() . '"', $startSkeleton);
-        $startSkeleton = preg_replace('/pageedit_priority=\"(.*)\"/', 'pageedit_priority="' . $formData->getPriority() . '"', $startSkeleton);
-
-        //Concatenate skeleton + metadata + content
-        $finalContent = $startSkeleton . "\n" . $formData->getContent() . "\n\t\t" . $endSkeleton;
-
-        //Archives old file if content or metadata are different
-        if ($fs->exists($filePath) && file_get_contents($filePath) !== $finalContent) {
-            $this->archiveFile($page, $userId);
-        }
-
-        //Writes new file
-        $newFilePath = $folderPath . '/' . $page . '.html.twig';
-        $fs->dumpFile($newFilePath, $finalContent);
-        $fs->chmod($newFilePath, 0770);
-
-        //Clears the cache otherwise changes will not be reflected
-        $fs->remove($this->getParameter('kernel.cache_dir') . '/../prod/twig');
     }
 }