+++ DependencyInjection/c975LPageEditExtension.php
@@ -22,6 +22,12 @@ class c975LPageEditExtension extends Extension
      */
     public function load(array $configs, ContainerBuilder $container)
     {
+        $loader = new YamlFileLoader(
+            $container,
+            new FileLocator(__DIR__.'/../Resources/config')
+        );
+        $loader->load('services.yml');
+
         $configuration = new Configuration();
         $processedConfig = $this->processConfiguration($configuration, $configs);